After 78b7101d570aa874cd748b0488ddeda15d7657df, the page's connection would of course be the duplicated connection which is a plain NMConnection that user edits are done on.  THat of course doesn't implement NMSettingsConnectionInterface, and thus can't respond to requests for secrets.
Instead, we have the CEPage subclasses for each setting type pass back
the name of the setting they want secrets for, if they need secrets,
and let the editor ask for the secrets since it has the original
connection which usually implements NMSettingsConnectionInterface.

By moving the addition of the page's GtkWidget to after page
initialization, we also fix system VPN connections, which were broken
because the VPN page's plugin-provided widget is only valid after
the page is initialized, but the editor expected it to be valid
right after requesting secrets.  In the case of system secrets, the
request is over D-Bus and is async, which meant that the call would
complete and the page would initalize long after the connection editor
had tried to add the page to the UI.

original bzr status:
revno: 1270
committer: Dan Williams <dcbw@redhat.com>
timestamp: Fri 2009-10-02 12:30:36 -0700
message: