editor: preserve user-scope VPN connection secrets when moving from system scope
When moving a connection between user and system scope, when the addition
of the connection to the new settings service is complete, it's scope has
already been changed to it's new settings service.  Thus when deleting it
from the old service and telling the VPN plugins to remove secrets, we
need to use the connection's *original* scope, not the current scope,
to ensure that we don't blow away the secrets we just wrote to the
keyring when moving a VPN connection from system to user scope.

original bzr status:
revno: 1733
author: cee1 <fykcee1@gmail.com>
committer: Dan Williams <dcbw@redhat.com>
timestamp: Wed 2010-10-20 16:21:05 -0500