applet: provide simple D-Bus API for GNOME Shell indicator (bgo #642503)
Since the Shell's indicator is lagging a bit in functionality due
to missing bindings for stuff like gnome-keyring expose a few small
dbus methods to help it out.  This is a *PRIVATE API* that is
not guaranteed to be stable.

[cleaned up, refactored, and permissions checking added by dcbw]

original bzr status:
revno: 1866
author: Giovanni Campagna <gcampagna@src.gnome.org>
committer: Dan Williams <dcbw@redhat.com>
timestamp: Wed 2011-03-02 14:27:09 -0600