The active WiMAX network provider name should be bold, but it wasn't.
This was due to a limitation in the mobile broadband menu item,
which only accounted for GSM/CDMA usage of the active item.  For
GSM/CDMA, the "active item" (ie the item just underneath the
device header) displays either the active connection, or if not
connected, the registered network name and signal strength.  If
not connected, the item insensitive because you need to pick a
specific connection/APN to activate.

With WiMAX, we can scan and get signal strength, so instead of
showing connections in the menu, we show actual NSPs (ie, WiMAX
Network Service Providers) determined from a scan.  We don't
have any registration nonsense because the WiMAX device doesn't
register with a network until it's told to do so, which happens
when the user picks a network to connect to.  But due to the
GSM/CDMA stuff above, the active item if connected wouldn't be
bold.

original bzr status:
revno: 1915
committer: Dan Williams <dcbw@redhat.com>
timestamp: Tue 2011-03-15 12:11:46 -0500
message: