gsm: add registration status notification
The current UI requires the user to mouse-over and reveal a tool
tip (or click an indicator) to discover the registration state
(aka polling). If the user inadvertently re-registers on a
roaming network, the operation is silent until the user decides
to check the applet for network status, possibly incurring
expensive roaming tariffs in the mean time.

So to prevent silent registration changes, alert the user with a
notification when the GSM registration status changes to home or
to roaming.

original bzr status:
revno: 2053
author: Andrew Bird <ajb@spheresystems.co.uk>
committer: Dan Williams <dcbw@redhat.com>
timestamp: Wed 2011-06-15 13:13:50 -0500