There were a number of problems here:
1) if not secrets got saved, no D-Bus reply would get sent back to the
caller due to mishandling of the tracking keyring requests; if no
keyring requests were made the save callback was not called

2) turns out we *do* want to save VPN secrets here; the Agent is the
thing that should be saving secrets, not the VPN plugins themselves,
because if somebody changes a connection using something other than
nm-connection-editor the VPN plugins wouldn't be called to save
secrets anyway.  That's what the agent is for: to save secrets that
processes that update connections send to NM which sends them on to
the agents.

3) 'get' secrets requests weren't canceled due to the same issue
as #1; with the fixup of tracking keyring requests this should now
happen

original bzr status:
revno: 2084
committer: Dan Williams <dcbw@redhat.com>
timestamp: Wed 2011-07-13 12:31:56 -0500
message: