cdma: enable registration state home/roaming notifications
The current UI requires the user to mouse-over and reveal a tool
tip (or click an indicator) to discover the registration state
(aka polling). If the user inadvertently re-registers on a
roaming network, the operation is silent until the user decides
to check the applet for network status, possibly incurring
expensive roaming tariffs in the mean time.

So to prevent silent registration changes, alert the user with a
notification when the CDMA registration status changes to home or
to roaming.

original bzr status:
revno: 2092
author: Andrew Bird <ajb@spheresystems.co.uk>
committer: Dan Williams <dcbw@redhat.com>
timestamp: Tue 2011-07-19 16:28:05 -0500