applet: always show the applet
This is a patch that has been carried in Ubuntu for a while now. Instead of
hiding the applet if NetworkManager isn't running, make use of the messages
already included that mention that it's not, and just always keep showing
the applet; just with a different menu if NM is unavailable.

original bzr status:
revno: 2764
author: Mathieu Trudel-Lapierre <mathieu-tl@ubuntu.com>
committer: Dan Williams <dcbw@redhat.com>
timestamp: Thu 2013-03-07 15:59:37 -0600