This applies to EAP methods PEAP, TLS, TTLS.
When the user configures a connection without CA certificate,
previously a warning dialog was shown where the user could
either return to the connection configuration or proceed.

Change the UI so that the configuration mask has a checkbox
"No CA is required" and enable the "Connect" button only when
either a certificate is selected or the user choose to ignore
it.

The flag no_ca_required is saved in the gsettings of the user.
When creating a new connection, the checkbox is always unchecked, so
that the user must make an active decision. When editing connections,
the flag is only set, if there is no ca cert filename and the
ca_cert_ignored flag is not set.

The nag dialog is still here and active. The next commit will
remove it.

https://bugzilla.redhat.com/show_bug.cgi?id=758076
https://bugzilla.redhat.com/show_bug.cgi?id=809489

Signed-off-by: Thomas Haller <thaller@redhat.com>

original bzr status:
revno: 2889
committer: Thomas Haller <thaller@redhat.com>
timestamp: Tue 2013-08-27 23:10:42 +0200
message: