The commit 3cae06120a4dd9bebd5f2e57e181ea95e87c8ac4 fixed the issue of loading the username/password when editing a connection with EAP-PEAP or EAP-TTLS (using EAPMethodSimple) or EAP-LEAP (using EAPMethodLEAP).
There was a small issue when you edited such a connection: when opening
the configuration dialog, the 'Apply' button was disabled until you switched
to the security tab.

The problem was, that in order to enable the 'Apply' button, 'validate' of
the security dialog will be called. 'validate' looks at the content of the UI
fields whether username and password are set. Due to a bug, they were
unset until 'widget_realize' was called, i.e. not before you switched to the
security tab the first time.

Related to:
https://bugzilla.redhat.com/show_bug.cgi?id=1000564

Signed-off-by: Thomas Haller <thaller@redhat.com>

original bzr status:
revno: 2906
committer: Thomas Haller <thaller@redhat.com>
timestamp: Wed 2013-09-11 16:05:55 +0200
message: