Do not allow to input unexpected (ascii) characters as DNS servers, both for IPv4 and IPv6.
This especially disallows to add a %scope-id suffix for IPv6 link-local
addresses (because '%' is not allowed). The reason to disallow this is
that link-local addresses are always scoped to the connecting device
anyway, so this parameter would get ignored anyway.

This is related to bug
https://bugzilla.redhat.com/show_bug.cgi?id=962449

Signed-off-by: Thomas Haller <thaller@redhat.com>

original bzr status:
revno: 2926
committer: Thomas Haller <thaller@redhat.com>
timestamp: Fri 2013-10-04 19:30:59 +0200
message: