The network type was never used with ModemManager 1.0+, and NM itself hasn't supported anything except ModemManager 1.0+ since 0.9.8.
Changing the network type is often a very heavy operation for the
modem, which may involve re-initialization or hardware reset, and
is not an appropriate operation to perform when connecting to
a network.  It should be an out-of-band action.

original bzr status:
revno: 3125
committer: Dan Williams <dcbw@redhat.com>
timestamp: Wed 2015-01-14 10:05:43 -0600
message: