If the connection is invalid, the "Connect" button in the Wi-Fi password dialog stays disabled and makes it hard for the user to understand the reason.
Especially when connecting to certain EAP-typed Wi-Fi connections, the
connection might have no CA certificate set. However, if the connection
was not previously edited with nm-connection-editor, the connection is
considered invalid by nm-applet because "No CA certificate is required"
is unchecked.
As this flag is only stored in the gsetting of the user, NetworkManager
and nm-applet disagree about whether the connection is valid.

Add a tooltip to the connect button to indicate to the user that the
connection must be modified first.

Steps to reproduce:

  - create a Wi-Fi connection of type EAP (PEAP, TLS, or TTLS) outside
    of nm-applet/nm-connection-editor. Or alternatively, clear the
    gsettings entry with:
       dconf reset -f /org/gnome/nm-applet/eap/
  - try to connect to the Wi-Fi. Note that the "Connect" button of the
    password dialog is diabled (just like the "Save" button in nm-ce).

https://bugzilla.gnome.org/show_bug.cgi?id=754172
https://bugzilla.redhat.com/show_bug.cgi?id=1247885

original bzr status:
revno: 3279
committer: Thomas Haller <thaller@redhat.com>
timestamp: Wed 2015-09-02 13:31:39 +0200
message: